requirements.txt
fastapi~=0.115.5
python-dotenv~=1.0.1
SQLAlchemy~=2.0.36
alembic~=1.14.0
pydantic~=2.10.2
pandas~=2.2.3